Move body parser to its own package
