+++ cmd/test.go
@@ -8,6 +8,8 @@ import (
 	"net/url"
 	"strings"
 
+	"github.com/ChrisMcKenzie/accord/pkg/parser"
+
 	accord "github.com/ChrisMcKenzie/accord/pkg"
 	"github.com/ChrisMcKenzie/accord/pkg/httptest"
 	"github.com/fatih/color"
@@ -32,14 +34,6 @@ var testCmd = &cobra.Command{
 
 		test(args[0])
 	},
-}
-
-type byteBufferReadCloser struct {
-	bytes.Buffer
-}
-
-func (b *byteBufferReadCloser) Close() error {
-	return nil
 }
 
 func init() {
@@ -68,9 +62,9 @@ func server(host, uri string, query map[string]string) *url.URL {
 
 func test(host string) {
 	ctx.ProcessEndpoints(func(ep *accord.Endpoint) {
-		var buf byteBufferReadCloser
+		var buf parser.ByteBufferReadCloser
 		if ep.Request != nil {
-			parser := Parser{Headers: ep.Request.Headers, Body: ep.Request.Body}
+			parser := parser.Parser{Headers: ep.Request.Headers, Body: ep.Request.Body}
 			buf, _ = parser.Parse()
 		}
 
@@ -104,7 +98,7 @@ func compareResponse(resp *http.Response, expect *accord.Response) error {
 		return err
 	}
 
-	parser := Parser{Headers: expect.Headers, Body: expect.Body}
+	parser := parser.Parser{Headers: expect.Headers, Body: expect.Body}
 	respBody, _ := parser.Parse()
 	if body.String() != respBody.String() {
 		diff := difflib.ContextDiff{
