+++ pkg/parser/parser_test.go
@@ -1,4 +1,4 @@
-package cmd
+package parser
 
 import (
 	"net/http"
